TEMPORARY: riscv: fu540-c000: add CLINT and ndev data for the BBL
The current BBL code needs CLINT data and the number of devices
possible to connect to the PLIC to be present in the DT data that it's
passed, otherwise it won't boot.  Linux doesn't need either of these,
so this is a temporary patch that is not meant for upstreaming.