rtc: Add driver for Synopsys Designware APB RTC
Designware APB is a programmable RTC compliant with AMBA 2.0
Advanced Peripheral Bus(APB).

Basic support for read & write time, read & set alarm, and
interrupt has been added. It has a programmable binary counter which
can be defined by the user. For now, we are using a pre-scaler counter
value of 32768(RTC XO) to get a resolution of 1Hz.

Signed-off-by: Himanshu Jha <himanshu.jha@sifive.com>