riscv: nb2:i2c: designware: dw_i2c_init_driver as module init
In fpga0p1 clock module(PRCI) is not available.So using a fixed-clock
as a clock source which is getting probed after i2c probing and result of
that i2c probing was failing.
To resolved this issue i2c bus subsys_initcall initialization
replace with module init for NB2_SOC platform.