Improvements to aes encryption logic
Some improvements are added after testing the aes code.

Successfully tested aes encrypt/decrypt for following modes:
CBC
ECB
CTR

All combinations are tested (DMA/Non-DMA and Interrupt/Polling)

Due to openssl limitations, could not test CFB and OFB modes but it is
expected to work as the logic is same as the above mentioned modes.